Default to stable2 patch since 3.14 is EOL
It was announced in https://lkml.org/lkml/2016/9/11/28 that the 3.14 kernel is
at EOL, will receive no more updates, and should not be used any longer. The 4.4
kernel, with correspoding "stable2" patches should now be defaulted to, as it is
now supposed to be the most stable LTS kernel that is still receiving updates.

Also, bumps SD version to latest.