Supports building under libvirt provider
The previous config assumes that VirtualBox is used as the Vagrant
backend, and accordingly bumped the specs (RAM and CPU count) only under
VirtualBox. Although the previous config still created the `grsec-build`
host under the vagrant-libvirt backend when explicitly specified, the
host was created with only 512MB of RAM, which is insufficient for
building a kernel.

The existing logic for dynamically determining the max CPU count on the
host machine is now applied to libvirt as well as virtualbox for the
backend provider. Also bumping RAM to 2048MB, same as under virtualbox.