Apply pax flags to grub-script-check binary
The only direct call by any hook to any grub utility during the kernel install
process is to `/etc/kernel/postinst.d/zz-update-grub`. `zz-update-grub` is a
shell script that calls a shell script that calls another shell script that may
call the following binaries: `grub-probe`, `grub-file`, `grub-editenv`, and
`grub-script-check`. Consulting all documentation and live examples (including
the grsec wiki, and paxrat and Arch's linux-pax-flags config files), it seems
that to minimize the chance of errors on a potentially very diverse set of
machines we will want to set `paxctl -zCE` on `/usr/bin/grub-script-check` even
though as far as we know the lack of this flag has not caused any errors for SD
users. I did not see any flags set on `grub-file` or `grub-editenv` anywhere,
and we're already setting flags on the other binaries.