Makes Version field in built packages more explicit
Previously was lazily using the `10.00-manual` string, which makes
subsequent upgrades painful, since the package "Version" strings don't
change. One solution is to use the metapackage to abstract the version
dependencies away, but that's just sweeping dust under the rug.

Since the entire build command is declared as a default role var, users
of the role can still override this setting at will.

Closes #76.