Smarter logic for determining whether connection is local
The previous implementation enabled or disabled reboots based on the
value of `ansible_connection`, specifically whether it was set to
'local', as it would be when running against localhost. This worked in
nearly all cases where the var was used in the role, *except* for the
post-reboot `wait_for` task, which was a `local_action`.

In the context of a `local_action` task, Ansible sets
ansible_connection=local automatically, causing the wait_for task to be
skipped even when the role was applied to remote hosts. By setting the
var via `set_fact`, it will persist even in the context of local_action.